NewLib: - Duplicate functionality files removed; - Refactoring of file handling functions; - Removed broken impliments. Gears (C + TinyGL): - Removed because it duplicates an existing example on Fasm and uses unsupported wrappers on the KOS API. KosJS: - Removed. The MuJS port is too old and not used anywhere. Support is not profitable. Backy: - Removed useless GCC version. Support is not profitable. DGen-SDL and SQLite3 - Fix after removing broken "dirent.h". Fridge: - Moving the KOS API wrapper to avoid compilation errors. Udis86, uARM and 8086tiny: - Fix after removing redundant "kos_LoadConsole.h".
git-svn-id: svn://kolibrios.org@9952 a494cfbc-eb01-0410-851d-a64ba20cac60

+++ contrib/sdk/sources/newlib/libc/sys/read.c
@@ -29,7 +29,7 @@ ssize_t read(int fd, void *buf, size_t cnt)
     char     *buffer = buf;
     int       rc;
     int       h;
-    unsigned amount_read;
+    size_t amount_read;
     int err;
 
     __io_handle *ioh;
